api/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
stripe==11.2.0
docker==7.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
email-validator==2.2.0

